requirements.txt
pandas
notebook
numpy
matplotlib
seaborn
pyYAML
tqdm
path
types-PyYAML
openai
langchain
streamlit
python-dotenv
PyPDF2

-e .